implement routes for mvp
