--- server/routes/fruits.ts
@@ -1,46 +1,0 @@
-import { Router } from 'express'
-import checkJwt, { JwtRequest } from '../auth0.ts'
-import { StatusCodes } from 'http-status-codes'
-
-import * as db from '../db/scores.ts'
-
-const router = Router()
-
-router.get('/', async (req, res) => {
-  try {
-    const fruits = await db.getAllFruits()
-
-    res.json({ fruits: fruits.map((fruit) => fruit.name) })
-  } catch (error) {
-    console.log(error)
-    res.status(500).json({ message: 'Something went wrong' })
-  }
-})
-
-router.get('/:id', async (req, res, next) => {
-  try {
-    const fruit = await db.getFruitById(req.params.id)
-    res.json(fruit)
-  } catch (err) {
-    next(err)
-  }
-})
-
-router.post('/', checkJwt, async (req: JwtRequest, res, next) => {
-  if (!req.auth?.sub) {
-    res.sendStatus(StatusCodes.UNAUTHORIZED)
-    return
-  }
-
-  try {
-    const { owner, name } = req.body
-    const id = await db.addFruit({ owner, name })
-    res
-      .setHeader('Location', `${req.baseUrl}/${id}`)
-      .sendStatus(StatusCodes.CREATED)
-  } catch (err) {
-    next(err)
-  }
-})
-
-export default router
