+++ server/server.ts
@@ -1,13 +1,15 @@
 import express from 'express'
 import * as Path from 'node:path'
 
-import fruitRoutes from './routes/fruits.ts'
+import cryptids from './routes/cryptids.ts'
+import scores from './routes/scores.ts'
 
 const server = express()
 
 server.use(express.json())
 
-server.use('/api/v1/fruits', fruitRoutes)
+server.use('/api/v1/cryptids', cryptids)
+server.use('/api/v1/scores', scores)
 
 if (process.env.NODE_ENV === 'production') {
   server.use(express.static(Path.resolve('public')))
